added contacts section to the tails page, move tails to the separate component
